requirements.txt
alabaster==0.7.12
appdirs==1.4.3
arrow==0.10.0
asn1crypto==0.24.0
attrs==18.2.0
autobahn==0.15.0
Babel==2.6.0
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.0
certifi==2018.11.29
cffi==1.7.0
chardet==3.0.4
click==6.7
colorama==0.3.7
configparser==3.7.1
coverage==4.1
cryptography==2.4.2
docker==2.5.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2; python_version < "3.2"
future==0.16.0
gitdb2==2.0.5
GitPython==2.1.5
idna==2.8
imagesize==1.1.0
incremental==17.5.0
ipaddress==1.0.16
Jinja2==2.10.1
jsl==0.2.4
jsonpatch==1.15
jsonpointer==2.0
jsonschema==2.6.0
klein==16.12.0
MarkupSafe==1.1.0
mock==2.0.0
msgpack-python==0.5.6
netifaces==0.10.4
nose==1.3.7
pbr==1.10.0
psutil==4.3.0
pulsectl==17.12.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.16
pycurl==7.43.0
pyflakes==1.6.0
Pygments==2.3.0
PyJWT==1.5.3
pyOpenSSL==17.5.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==5.1
requests==2.21.0
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.60
service-identity==16.0.0
six==1.10.0
smmap2==2.0.5
smokesignal==0.7.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
Twisted==16.2.0
txaio==2.5.1
txdbus==1.0.13
txmsgpackrpc==1.1
typing==3.6.6
urllib3==1.24.2
websocket-client==0.37.0
Werkzeug==0.14.1
zope.interface==4.2.0
